refactor : image/application 폴더 밑으로 이동
