+++ src/main/java/com/kakaotech/team14backend/post/application/PostService.java
@@ -1,7 +1,7 @@
 package com.kakaotech.team14backend.post.application;
 
 import com.kakaotech.team14backend.image.domain.Image;
-import com.kakaotech.team14backend.inner.image.usecase.CreateImage;
+import com.kakaotech.team14backend.image.application.CreateImage;
 import com.kakaotech.team14backend.inner.point.model.GetPointPolicy;
 import com.kakaotech.team14backend.inner.point.usecase.GetPointUsecase;
 import com.kakaotech.team14backend.inner.post.usecase.FindMyPostUsecase;
